move sync metadata table to autofill db (from secure store)
