+++ autofill/autofill-impl/build.gradle
@@ -55,6 +55,9 @@ dependencies {
     implementation AndroidX.work.runtimeKtx
     implementation AndroidX.room.ktx
 
+    implementation "com.jakewharton.threetenabp:threetenabp:_"
+    testImplementation "org.threeten:threetenbp:_"
+
     // Testing dependencies
     testImplementation "org.mockito.kotlin:mockito-kotlin:_"
     testImplementation Testing.junit4
