+++ autofill/autofill-impl/src/test/java/com/duckduckgo/autofill/impl/sync/SyncLoginCredentialsTest.kt
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.securestorage.store.sync
+package com.duckduckgo.autofill.impl.sync
 
 import androidx.room.Room
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.duckduckgo.app.CoroutineTestRule
-import com.duckduckgo.securestorage.store.db.SecureStorageDatabase
+import com.duckduckgo.autofill.store.AutofillDatabase
+import com.duckduckgo.autofill.store.sync.LoginCredentialsSync
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
@@ -32,40 +33,33 @@ import org.robolectric.RuntimeEnvironment
 @RunWith(AndroidJUnit4::class)
 internal class SyncLoginCredentialsTest {
 
-    @get:Rule
-    @Suppress("unused")
-    val coroutineRule = CoroutineTestRule()
+    @get:Rule @Suppress("unused") val coroutineRule = CoroutineTestRule()
 
-    private val db = Room.inMemoryDatabaseBuilder(RuntimeEnvironment.getApplication(), SecureStorageDatabase::class.java)
-        .allowMainThreadQueries()
-        .build()
+    private val db = Room.inMemoryDatabaseBuilder(RuntimeEnvironment.getApplication(), AutofillDatabase::class.java).allowMainThreadQueries().build()
     private val dao = db.syncLoginCredentialsDao()
 
     private val testee = SyncLoginCredentials(dao)
 
-    @Test
-    fun whenLoginIdNotFoundThenReturnNewSyncId() {
+    @Test fun whenLoginIdNotFoundThenReturnNewSyncId() {
         val syncId = testee.getSyncId(123L)
 
         assertNotNull(syncId)
     }
 
-    @Test
-    fun whenLoginIdExistsThenReturnSyncId() {
+    @Test fun whenLoginIdExistsThenReturnSyncId() {
         val loginId = 123L
         val syncId = "syncId"
-        dao.insert(LoginCredentialsSync(syncId = syncId, id= loginId))
+        dao.insert(LoginCredentialsSync(syncId = syncId, id = loginId))
 
         val result = testee.getSyncId(loginId)
 
         assertEquals(syncId, result)
     }
 
-    @Test
-    fun whenEntityRemovedThenDeleteAtUpdated() {
+    @Test fun whenEntityRemovedThenDeleteAtUpdated() {
         val loginId = 123L
         val syncId = "syncId"
-        dao.insert(LoginCredentialsSync(syncId = syncId, id= loginId))
+        dao.insert(LoginCredentialsSync(syncId = syncId, id = loginId))
 
         testee.onEntityRemoved(loginId)
