+++ secure-storage/secure-storage-store/build.gradle
@@ -38,8 +38,6 @@ dependencies {
     implementation AndroidX.room.ktx
     implementation AndroidX.security.crypto
     implementation Square.okio
-    implementation project(path: ':common')
-    implementation "org.threeten:threetenbp:_"
 
     kapt AndroidX.room.compiler
